+++ lib/sigepweb.dart
@@ -2,5 +2,6 @@ library sigepweb;
 
 export 'src/core.dart';
 export 'src/utils.dart';
+export 'src/constants.dart';
 export 'src/preco_prazo.dart';
 export 'src/models/contrato.dart';