+++ lib/src/models/calc_preco_prazo_item.dart
@@ -1,6 +1,8 @@
 import 'package:sigepweb/sigepweb.dart';
+import 'package:sigepweb/src/constants.dart';
 
 class CalcPrecoPrazoItem {
+  String nome;
   String codigo;
   double valor;
   int prazoEntrega;
@@ -15,7 +17,7 @@ class CalcPrecoPrazoItem {
   Map obsFim;
 
   CalcPrecoPrazoItem.fromJson(Map<String, dynamic> json)
-      : codigo = json['Codigo']['\$t'],
+      : codigo = json['Codigo']['\$t']?.padLeft(5, '0'),
         valor = SgUtils.toDouble(json['Valor']['\$t']),
         prazoEntrega = int.tryParse(json['PrazoEntrega']['\$t']) ?? 0,
         valorMaoPropria = SgUtils.toDouble(json['ValorMaoPropria']['\$t']),
@@ -28,5 +30,9 @@ class CalcPrecoPrazoItem {
         msgErro = json['MsgErro']['\$t'],
         valorSemAdicionais =
             SgUtils.toDouble(json['ValorSemAdicionais']['\$t']),
-        obsFim = json['obsFim']['\$t'];
+        obsFim = json['obsFim']['\$t'] {
+    //
+    // with the code we set the name of the service
+    nome = ServicosPostagem.nomeServico[codigo] ?? 'Correios';
+  }
 }