+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # 0.0.1
 
-* Initializing the package only with CalcPrecoPrazo method available
+* Initializing the package only with CalcPrecoPrazo method available.
+
+# 0.0.2
+
+* Name of the service with model CalcPrecoPrazoItem and method for test CEP numbers.